obtain latex file structure from datafile
Before this, the latex file structure was obtained from the dicts.py module at every level.

However, this module does not reflect some changes that are made to the data set later, and hence, now the latex files are made from the datafiles themselves. The scripts now look for the latest version of the data files, and create a tablewith coordinates and variables separated.