rename flags and tests to new names + sonde_id as dim
The manuscript for JOANNE revealed that the original names for the flags and tests (ind, srf) were not intuitive and a change in the naming scheme would help with better understanding of the QC tests. Therefore, the variables have been renamed in the status dataset. Note that before the dataset is created, i.e. when the actual QC tests are running, the original names have been maintained in the script. The names are changed just before saving the status dataset, and are thus only to be seen as a quick fix.