include data doi and software doi in dataset
The data doi has been obtained from AERIS and reserved. Similarly, the doi for the software has been obtained and reserved from Zenodo. These information have been pre-entered into the joanne __init__.py file, so as to keep the respective persistent identifiers within the software and datasets. For now, a prefix "NOTREADY-" has been added, so as to not have any other copies of the software or dataset with the same doi information. This prefix will be removed before the final run, for which the dois have been reserved.