include reference study in the global attrs
The JOANNE manuscript to be published in the ESSD journal does not have a doi yet, and therefore only an unpublished citation is provided as value for the "reference" global attribute for all files starting from Level-2.

Level-4 initially had a reference study (Bony & Stevens, 2019, Measuring Area-Averaged Vertical Motions with Dropsondes, JAS) but since the dataset itself has a manuscript now, the aforementioned study is no longer needed and the current reference works as a more relevant source.