remove check of max version for L3
This is a temporary fix to a problem where importing this particu
lar module raised an error that max(vers) was empty because the f
ile directory and file names are all hardcoded. A future JOANNE u
pdate should fix this with paths not being part of src file, but
rather a separate config file. But this particular change does no
t need the aforementioned changes in how paths are provided.

TLDR: L3 directory should contain only one L3 file; more than one
 will throw an error