remove READMEs from source folders
The README files within all directories from Level-0 to Level-4
contained outdated information. These were mainly to inform the users about
the data structure when the dataset was still in the process of creation.
Now that JOANNE has its own publication out describing the dataset, these READMEs
are not needed, are outdated and maintaining them only creates points of failure
in documentation, which is already redundant because of the aforementoned publication.